Re-work a step.
One problem is that
    the Number value of X &times; Y
is ambiguous -- it could mean:
    (the Number value of X) &times; Y
or:
    the Number value of (X &times; Y)

Another problem is that _n_ is mathematical,
so in 10<sup>_n_</sup>, 10 should be mathematical too.
And then, depending on how the above ambiguity resolves,
you'd have to either convert that to a Number
or change &times; to a mathematical operator.

It's simpler (I think) to do all the arithmetic in the math realm,
and then convert just the result to a Number.

(I'm assuming that 'plus' and 'times' as *words*
are mathematical operators.)