Layering: Rephrase as HostEnqueuePromiseJob
Rather than using a string queue name, this patch uses separate hooks
for separate things that may be scheduled, in order to permit the
clear expression of invariants which differ based on the particular
type of job. For example, Promise Jobs are required to be served in
FIFO order, whereas WeakRef Jobs are unordered.