Tweak some of the wording around jobs
... mainly to bring agents into it more.

(Also delete 2 'es6num' lines.)

Note that the sentence
"Only one Job may be actively undergoing evaluation at any point in time."
is redundant given the run-to-completion requirement.

(In fact, even the run-to-completion sentence is pretty much redundant
given the requirement to start when the execution context stack is empty.
But I think it's useful.)