Replace HostEnqueuePromiseJob with HostEnqueueJob
It's simpler and I think it reads better.

The major benefit would be when you can add a new "job group"
with as little as one line.

I called it 'jobGroup' (rather than the analogous 'queueName' parameter
of EnqueueJob in the status quo spec) because I don't think
the spec can dictate the queues that the agent must have.

Note that HostEnqueueJob's parent clause is basically redundant:
everything it says is also conveyed by the HostEnqueueJob clause.
But I left it there in case it becomes more of a useful overview.